拖拽完成190215

+++ application/admin/controller/User.php
@@ -8,7 +8,9 @@ class User extends \app\common\controller\BaseController
     public function data_grid($page = 1, $rows = 20, $sort = 'id', $order = 'desc') {
         if ($this->request->isAjax()) {
             $count = db('user')->count();
-            $list = UserModel::with('staff')->order([$sort => $order])->limit($rows)->page($page)->select();
+            
+            $list = UserModel::with('groups')->order([$sort => $order])->limit($rows)->page($page)->select();
+            //dump($list);
             return ['total' => $count, 'rows' => $list ? $list : ''];
         }
     }
@@ -16,23 +18,17 @@ class User extends \app\common\controller\BaseController
     public function create() {
         if ($this->request->isPost()) {
             $data = input('post.');
-            $data['create_time'] = get_time();
+            //$data['create_time'] = get_time();
             $data['password'] = md5($data['password']);
-            $result = 0;
             // 启动事务
             Db::startTrans();
             try {
                 $userId = Db::name('user')->strict(false)->insertGetId($data);
-                $staff = Db::name('staff')->find($data['staff_id']);
-                if($staff){
-                    $result1 = Db::name('staff')->update(['id' => $data['staff_id'], 'user_id' => $userId]);
-                    //新增用户权限分配
-                    $result2 = Db::name('auth_group_access')->insert([
-                        'uid' => $userId,
-                        'group_id' => $staff['group_id']
-                    ]);
-                }
-                $result = $result1 && $result2;
+                //新增用户权限分配
+                $result = Db::name('auth_group_access')->insert([
+                    'uid' => $userId,
+                    'group_id' => $data['group_id']
+                ]);
                 // 提交事务
                 Db::commit();
             } catch (\Exception $e) {
@@ -51,16 +47,24 @@ class User extends \app\common\controller\BaseController
 
     public function detail() {
         $id = input('id/d', 0);
-        $model = db('user')->find($id);
-        return json($model);
+        $user = UserModel::get($id);
+        $groups = $user->groups;
+        foreach ($groups as $group) {
+            // 输出用户的角色名
+            $user['group_id'] = $group->id;
+        }
+        return json($user);
     }
 
     public function delete() {
         $ids = input('ids', '');
+        $arr = explode(',', $ids);
+        if(in_array('1',$arr)){
+            return ['Success' => false, 'Message' => '管理员不能删除！'];
+        }
         // 启动事务
         Db::startTrans();
         try {
-            Db::name('staff')->where('user_id', 'in', $ids)->setField('user_id', 0);
             // 根据主键删除
             Db::name('user')->delete($ids);
             Db::name('auth_group_access')->where('uid', 'in', $ids)->delete();
@@ -81,21 +85,15 @@ class User extends \app\common\controller\BaseController
             // 启动事务
             Db::startTrans();
             try {
-                $user = Db::name('user')->find($data['id']);
-                Db::name('staff')->update(['id' => $user['staff_id'], 'user_id' => 0]);
-                
+                //$user = Db::name('user')->find($data['id']);
                 Db::name('auth_group_access')->where('uid', $data['id'])->delete();
                 
-                $result = Db::name('user')->strict(false)->update($data);
-                $staff = Db::name('staff')->find($data['staff_id']);
-                if($staff){
-                    $result = Db::name('staff')->update(['id' => $data['staff_id'], 'user_id' => $data['id']]);
-                    //新增用户权限分配
-                    $result2 = Db::name('auth_group_access')->insert([
-                        'uid' => $user['id'],
-                        'group_id' => $staff['group_id']
+                Db::name('user')->strict(false)->update($data);
+                $result = Db::name('auth_group_access')->insert([
+                        'uid' => $data['id'],
+                        'group_id' => $data['group_id']
                     ]);
-                }
+                
                 // 提交事务
                 Db::commit();
             } catch (\Exception $e) {
@@ -113,15 +111,4 @@ class User extends \app\common\controller\BaseController
             return view();
         }
     }
-
-//    private function insert_test(){
-//        $data = [];
-//        for($i = 1; $i <= 30; $i++){
-//            $user['accounts'] = 'test'.$i;
-//            $user['password'] = 'e10adc3949ba59abbe56e057f20f883e';
-//            $user['state'] = '正常';
-//            $data[] = $user;
-//        }
-//        db('user')->insertAll($data);
-//    }
 }
